pyproject.toml
[project]
name = "fx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastai>=2.7.13",
  "face-recognition>=1.3.0",
  "pillow>=10.2.0",
  "requests>=2.31.0",
  "numpy>=1.26.0",
]
